Added option to `overload` miriad parameters
Added an option to overload miriad keys and values when creating the mirstr

This, I hope, would be useful when you have a imaging/calibration pipeline and wish to overwrite specific values from some `base` pipeline step. Consider for example replacing the interval used in the third round of selfcal in an imaging pipeline. Rather then having to either rewrite/duplicate an entire pipeline function or include potentially dozens of other default keyword arguments, can pass a single dict with items to update or add